+++ .kurtosis/kurtosis.sh
@@ -1,7 +1,9 @@
 #!/usr/bin/env bash
 
-# Copyright (c) 2020 - present Kurtosis Technologies LLC.
+#
+# Copyright (c) 2021 - present Kurtosis Technologies Inc.
 # All Rights Reserved.
+#
 
 # !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! WARNING !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 #
@@ -20,14 +22,15 @@ set -euo pipefail
 # Can make this configurable if needed
 KURTOSIS_DIRPATH="${HOME}/.kurtosis"
 
-KURTOSIS_CORE_TAG="1.16"
+KURTOSIS_CORE_TAG="1.18"
 KURTOSIS_DOCKERHUB_ORG="kurtosistech"
 INITIALIZER_IMAGE="${KURTOSIS_DOCKERHUB_ORG}/kurtosis-core_initializer:${KURTOSIS_CORE_TAG}"
 API_IMAGE="${KURTOSIS_DOCKERHUB_ORG}/kurtosis-core_api:${KURTOSIS_CORE_TAG}"
 
 POSITIONAL_ARG_DEFINITION_FRAGMENTS=2
 
-
+# The identifier that will be prefixed to all objects run in the Kurtosis testing framework
+KURTOSIS_TESTING_IDENTIFIER="KTT"
 
 # ============================================================================================
 #                                      Arg Parsing
@@ -198,31 +201,14 @@ else
     echo "Successfully pulled latest version of API image"
 fi
 
-# Kurtosis needs a Docker volume to store its execution data in
-# To learn more about volumes, see: https://docs.docker.com/storage/volumes/
-sanitized_image="$(echo "${test_suite_image}" | sed 's/[^a-zA-Z0-9_.-]/_/g')"
-suite_execution_volume="$(date +%Y-%m-%dT%H.%M.%S)_${sanitized_image}"
-if ! docker volume create "${suite_execution_volume}" > /dev/null; then
-    echo "ERROR: Failed to create a Docker volume to store the execution files in" >&2
-    exit 1
-fi
-
 if ! mkdir -p "${KURTOSIS_DIRPATH}"; then
     echo "ERROR: Failed to create the Kurtosis directory at '${KURTOSIS_DIRPATH}'" >&2
     exit 1
 fi
 
-if ! execution_uuid="$(uuidgen)"; then
-    echo "ERROR: Failed to generate a UUID for identifying this run of Kurtosis" >&2
-    exit 1
-fi
-if ! execution_uuid="$(echo "${execution_uuid}" | tr '[:lower:]' '[:upper:]')"; then
-    echo "ERROR: Failed to uppercase execution instance UUID" >&2
-    exit 1
-fi
-
+execution_id="${KURTOSIS_TESTING_IDENTIFIER}$(date +%FT%H.%M.%S)-${RANDOM}"
 docker run \
-    --name "${execution_uuid}__initializer" \
+    --name "${execution_id}__initializer" \
     \
     `# The Kurtosis initializer runs inside a Docker container, but needs to access to the Docker engine; this is how to do it` \
     `# For more info, see the bottom of: http://jpetazzo.github.io/2015/09/03/do-not-use-docker-in-docker-for-ci/` \
@@ -231,9 +217,6 @@ docker run \
     `# Because the Kurtosis initializer runs inside Docker but needs to persist & read files on the host filesystem between execution,` \
     `#  the container expects the Kurtosis directory to be bind-mounted at the special "/kurtosis" path` \
     --mount "type=bind,source=${KURTOSIS_DIRPATH},target=/kurtosis" \
-    \
-    `# The Kurtosis initializer image requires the volume for storing suite execution data to be mounted at the special "/suite-execution" path` \
-    --mount "type=volume,source=${suite_execution_volume},target=/suite-execution" \
     \
     `# The initializer container needs to access the host machine, so it can test for free ports` \
     `# The host machine's IP is available at 'host.docker.internal' in Docker for Windows & Mac by default, but in Linux we need to add this flag to enable it` \
@@ -245,12 +228,11 @@ docker run \
     --env CLIENT_SECRET="${client_secret}" \
     --env CUSTOM_PARAMS_JSON="${custom_params_json}" \
     --env DO_LIST="${do_list}" \
-    --env EXECUTION_UUID="${execution_uuid}" \
+    --env EXECUTION_ID="${execution_id}" \
     --env IS_DEBUG_MODE="${is_debug_mode}" \
     --env KURTOSIS_API_IMAGE="${API_IMAGE}" \
     --env KURTOSIS_LOG_LEVEL="${kurtosis_log_level}" \
     --env PARALLELISM="${parallelism}" \
-    --env SUITE_EXECUTION_VOLUME="${suite_execution_volume}" \
     --env TEST_NAMES="${test_names}" \
     --env TEST_SUITE_IMAGE="${test_suite_image}" \
     --env TEST_SUITE_LOG_LEVEL="${test_suite_log_level}" \