+++ golang/testsuite/execution_impl/example_testsuite_configurator.go
@@ -1,14 +1,9 @@
-/*
- * Copyright (c) 2021 - present Kurtosis Technologies LLC.
- * All Rights Reserved.
- */
-
 package execution_impl
 
 import (
 	"encoding/json"
+	"github.com/kurtosis-tech/kurtosis-testsuite-starter-pack/golang/testsuite/testsuite_impl"
 	"github.com/kurtosis-tech/kurtosis-testsuite-api-lib/golang/lib/testsuite"
-	"github.com/kurtosis-tech/kurtosis-libs/golang/testsuite/testsuite_impl"
 	"github.com/palantir/stacktrace"
 	"github.com/sirupsen/logrus"
 	"strings"