+++ golang/testsuite/testsuite_impl/advanced_network_test/advanced_network_test_.go
@@ -1,13 +1,8 @@
-/*
- * Copyright (c) 2020 - present Kurtosis Technologies LLC.
- * All Rights Reserved.
- */
-
 package advanced_network_test
 
 import (
 	"github.com/kurtosis-tech/kurtosis-client/golang/lib/networks"
-	"github.com/kurtosis-tech/kurtosis-libs/golang/testsuite/networks_impl"
+	"github.com/kurtosis-tech/kurtosis-testsuite-starter-pack/golang/testsuite/networks_impl"
 	"github.com/kurtosis-tech/kurtosis-testsuite-api-lib/golang/lib/testsuite"
 	"github.com/palantir/stacktrace"
 	"github.com/sirupsen/logrus"
@@ -62,7 +57,7 @@ func (test *AdvancedNetworkTest) Run(network networks.Network) error {
 	}
 	logrus.Info("Incremented number of books read")
 
-	logrus.Info("Retrieving test person to verify number of books read person-retrieving API client...")
+	logrus.Info("Retrieving test person to verify number of books read by the person-retrieving API client...")
 	person, err := personRetrieverClient.GetPerson(testPersonId)
 	if err != nil {
 		return stacktrace.Propagate(err, "An error occurred getting the test person")