+++ typescript/testsuite/execution_impl/example_testsuite_args.ts
@@ -1,0 +1,4 @@
+export interface ExampleTestsuiteArgs {
+    readonly apiServiceImage: string;
+    readonly datastoreServiceImage: string;
+}